UL-96 - feat(DeletePhoto): Add UPopMenuItem in UILibrary and UIShowRoom
