+++ packages/ui_showroom/lib/ui_pages/ui_pages_export.dart
@@ -6,3 +6,4 @@ export 'src/onboarding_widgets/onboarding_widgets_export.dart';
 export 'src/qr_code/qr_code_export.dart';
 export 'src/u_user_profile_cards/u_user_profile_cards_export.dart';
 export 'src/dialog_widgets/dialog_widgets_export.dart';
+export 'src/pop_menu_item_widgets/pop_menu_item_widgets_export.dart';
